Refator: always flush full metrics

+++ cluster_benchmark/tests/benchmark/bench_cluster.rs
@@ -121,7 +121,7 @@ async fn do_bench(bench_config: &BenchConfig) -> anyhow::Result<()> {
         handles.push(h)
     }
 
-    leader.wait(timeout()).log_at_least(Some(total as u64), "commit all written logs").await?;
+    leader.wait(timeout()).log_at_least(Some(total), "commit all written logs").await?;
 
     let elapsed = now.elapsed();
